improve product identifier resolution
