+++ bundles/ru.arsysop.passage.lic.base/src/ru/arsysop/passage/lic/base/LicensingPaths.java
@@ -24,13 +24,10 @@ import java.net.URI;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-import ru.arsysop.passage.lic.registry.ProductDescriptor;
-
 public class LicensingPaths {
 
 	public static String FOLDER_LICENSING_BASE = ".passage"; //$NON-NLS-1$
 
-	public static String EXTENSION_LICENSE_DECRYPTED = ".lic"; //$NON-NLS-1$
 	public static String EXTENSION_LICENSE_ENCRYPTED = ".licen"; //$NON-NLS-1$
 	public static String EXTENSION_PRODUCT_PUBLIC = ".pub"; //$NON-NLS-1$
 	
@@ -43,19 +40,11 @@ public class LicensingPaths {
 
 	public static Path resolveConfigurationPath(String from, Object configuration) {
 		Path basePath = getBasePath(from);
-		String identifier = resolveProductIdentifier(configuration);
+		String identifier = LicensingConfigurations.resolveProductIdentifier(configuration);
 		if (identifier != null) {
 			return basePath.resolve(identifier);
 		}
 		return basePath;
-	}
-
-	public static String resolveProductIdentifier(Object configuration) {
-		if (configuration instanceof ProductDescriptor) {
-			ProductDescriptor product = (ProductDescriptor) configuration;
-			return product.getIdentifier();
-		}
-		return null;
 	}
 
 }
